refactor extractor
+ using handlers models
+ process every message in an tx
